Update codingguidelines.rst
Add most of the coding info from http://owncloud.org/contribute - this is where it belongs. Will link to it from owncloud.org ;-)

Of course if we agree it belongs on owncloud.org then we can make a dedicated page there. It at least won't fit on owncloud.org/contribute because of https://github.com/owncloud/www/issues/132