Create installing_apps.rst
I suggest to amend the documentation and add the following to the apps.owncloud.com website as well. Since this is a new section and describes issues under active discussion I am putting this up as an issue, inviting comments and corrections as well as instructions from the maintainers as to where exactly this might be placed in the documentation.