better handling of custom certificates
I followed your steps to setup webdav via command-line. I also use a custom certificate and thought, that it is a bad idea to not check the fingerprint of the certificate. I solved it, by addressing the "pem"-file in the configuration. I think this is a far better solution.

Regards