Dev Core Translation emphasize why not to split a sentence
While fixing a few minor grammar things, I also felt it would be useful to add an extra sentence to explain why translation does not work in little pieces of a sentence. I work in Nepal and speak Nepali as well as English. The Nepali grammar order is subject-object-verb and it does not work to translate "The dog" "bit" "the man" as 3 pieces - the result has to also be reordered as "The dog" "to the man" "bit". "the man" also has to get an object indicator on it, which would be different if it was the subject of the sentence...