Updated section on Apache with mod_fcgid with information on FcgidMaxRequestInMem
Quintessence of ownCloud issue #22036: FcgidMaxRequestInMem still needs to be significantly increased from its default value to avoid the occurence of segmentation faults when uploading big files, which is not a regular setting but serves as a workaround for Apache with mod_fcgid bug #51747. Cf. https://github.com/owncloud/core/issues/22036.