Updated with Nextcloud communication channels
Changed description to match that there are no longer Enterprise vs Free versions in Nextcloud, but Enterprise Support subscriptions Replaced doc.owncloud.org links with docs.nextcloud.org ones. Removed references to ownClouders Youtube channel as there doesn't seem to be a Nextcloud equivalent (because we're using the Nextcloud official channel for everything?) Not sure if we want to set up a "Nextclouders" YouTube channel or not. All links are active, except the iOS one, which still needs to be created.