Update from ownCloud to Nextcloud references
Updated tagline, description so it references a single code version with optional enterprise support and the links to the clients. This will need to be updated when we have a Nextcloud desktop client of our own. The iOS link is not yet a live page, but I plan to add it in a separate commit in the same PR.