Replace/remove link to Drone server
The main page for the Drone server of Nextcloud does not provide any
value to a regular user, as it is just a login page. Due to its use in
the Kanban document, its link was replaced by the webpage of the Drone
project itself; in the case of the codereviews document the link was
removed and replaced with a textual explanation about where to find the
results of each pull request build.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>