fix: change 'sqlite' to 'sqlite3'
On a fresh Nextcloud 13 Installation the default value for dbtype in config.php is 'sqlite3' (and not 'sqlite'). I also noticed that on previous versions. Despite the fact that Nextcloud 13 also works though, if that value is manually changed to 'sqlite', I would suggest to change that in the documentation to the actually used value created when installing Nextcloud - and that is 'sqlite3'.