Update source_installation.rst
Please rephrase my edit if necessary. 
Added a note that the nextcloud snap package version may differ from the current stable version. (currently it is stable nc12 and this is the nc13 documentation...)
It might also be wise to discuss with the user in advance that snap has certain drawbacks, such as being largely read only.