Update example_ubuntu.rst for Ubuntu 20.04 LTS
This documentation has been outdated since Ubuntu 18.04 LTS (mysql/mariadb don't prompt for a root password, so you need to create a Nextcloud user), but it doesn't work on Ubuntu 20.04 LTS because it uses explicit PHP version numbers and it omits some recommended PHP modules for Nextcloud Hub 18 and 19.  I've updated the instructions to run on Ubuntu 20.04 LTS and have tested each step.

This page now leaves a new Ubuntu 20.04 LTS system in a state to continue the "Installation on Linux" steps starting with configuring Apache2 without errors.

Signed-off-by: Nathan Haines <nhaines@ubuntu.com>