Update system_requirements.rst
A partner had some problems installing Nextcloud, at the end we figured out they created a db user but not the database as such because it was never mentioned in our documentation. Not sure what's the right place and if we should have more detailed documentation for different databases. For now I added it to the requirements as a additional bullet point