Update example_ubuntu.rst
I'm somewhat new to the Linux world and was doing this on a fresh install of Ubuntu. I needed to run the apt update command in order to fetch package information for the following commands and get them to run/install the necessary packages. 

I also needed to start the MySQL server instance before I was able to run the listed commands.