Update "Development Environment" guide
While setting up a local dev checkout of nextcloud/server, I noticed
that the part about creating the 'config' directory is outdated: this
directory already exists in the repository and doesn't need to be created
during setup.

I have fixed that part of the guide and also clarified some wordings
elsewhere in this guide.

Signed-off-by: Nina Pypchenko <22447785+nina-py@users.noreply.github.com>