Update PHP setting recommendations for php7
The recommended config for php.ini seems to be very outdated. This was mentioned in #4758 https://github.com/nextcloud/documentation/issues/4758
The user is told to install PDO_MySQL but the config given appears to be for an old version of the ext/mysql. It also contains comments `#` that are no longer allowed since php7.0
Also, at least in Ubuntu when you install the pdo_mysql it creates a working config with good defaults so it's confusing to a new user when reading the current document since they really wouldn't have to change anything on their setup.