Add note about when to enable SSL
As the manual currently reads it sounds like you should always enable SSL for your database. Added note explaining that is is only necessary when you connect remotely.
Since the default install for MariaDB or MySQL only allows communication on localhost through a unix socket, this will not need to be configured for most installs.