Fix missing image in "federated_cloud_sharing.rst"
This adds back the file removed in commit
8106b6002509fba38c238996179d8c5f85442f24 but still used in
"federated_cloud_sharing.rst".

The remote user shown in the image is
"username@sub.example.com/nextcloud", so the example name used in the
page did not match and thus was removed (it was not replaced as it felt
redundant with the "username" already used in the previous sentence).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>