Allow to serve static WebAssembly and TensorFlow Lite files
Since Talk 13 (and thus Nextcloud 23) WebAssembly (.wasm) and TensorFlow
Lite (.tflite) files need to be loaded from the web server to provide
certain features (like the background blur in the WebUI).

Those files can be treated in a similar way to other static resources,
and there should not be any problem caching or compressing them.
However, as compressed TensorFlow Lite files are only ~12% smaller,
the compression directive depends on the MIME type and there is no
standard MIME type for TensorFlow Lite files, for now only WebAssembly
files are compressed.

Depending on the setup "application/wasm" may not be associated with
".wasm" files, so the directive was added just in case, as the
compression is enabled through the MIME type and, besides that, the
browsers log a warning if the expected MIME type is not returned.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>