fix: Update links to calendar, contacts & sync_gnome doc (pim -> groupware)
Three links were redirecting to doc in the _pim_ directory which doesn't exist anymore and has been replaced with _groupware_.
Issues are encountered with the _firstrunwizard_ displaying links to the calendar and contacts doc :
- https://github.com/nextcloud/firstrunwizard/blob/master/templates/page.clients.php#L59
- https://github.com/nextcloud/firstrunwizard/blob/master/templates/personal-settings.php#L66

Also it seems that most translations are referring to the _pim_ directory even in translations for the _groupware_ one.
See : https://github.com/nextcloud/documentation/search?l=Gettext+Catalog&q=pim
Not being familiar with how the translation works here, an issue will be created to discuss it.

Signed-off-by: Am Br <reg.dev@am35.fr>