Add documentation for developping an existing app
The App development » Introduction documentation only discusses how to create a new app. Wanting to contribute to an existing app, I was confused by this. Once I figured out how to set everything up so that I could see my changes to the app in my development version of Nextcloud, I thought I would make this pull request to add this information to the docs.

Signed-off-by: Marcel Robitaille <mail@marcelrobitaille.me>