Add high level filesystem documention
From https://github.com/nextcloud/server/pull/26982

Adds some documentation for the filesystem layer, both a high level overview of how the various pieces interact and a high level guide for apps interacting with the filesystem.

Hopefully this will be useful to anyone trying to either use the filesystem or work on the filesystem.

Signed-off-by: Louis Chemineau <louis@chmn.me>