Add hint about "fileinfo" module being required for theming
The Theming app requires the "mime_content_type" function, which is part
of the "fileinfo" module, to identify the mime type of the custom
images.

Moreover, if the "imagick" module is installed and it has support for
SVG files a rasterized counterpart will be generated for the SVG app
icons. In this case "mime_content_type" is used to identify the mime
type of the original icon.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>