Update example_ubuntu.rst
I added in the MariaDB command to create a user the 'WITH GRANT OPTION'. I followed the tutorial and i wasn't able to connect to my database when i was in the install wizard on nextcloud. I deleted the database and user and start over with this solution i found online here: (https://help.nextcloud.com/t/trouble-installing-you-need-to-enter-details-of-an-existing-account/84657/9) Now all is well, I hope with this change to save a couple of hours to anybody who want to follow this tutorial.

Signed-off-by: Whurauk <10385898+Whurauk@users.noreply.github.com>