Use --option=value version for occ maintenance:install
The `--option=value` version of passing options to the `occ maintenance:install` command is now used in the documentation, instead of the `--option value` variant. This  solves issues with leading dashes in values, especially passwords. It also matches the `occ help maintenance:install` output.

Solves:
- https://github.com/nextcloud/server/issues/26109
- https://github.com/nextcloud/documentation/issues/8190

Furthermore values are now single quoted, which is important to avoid variable expansion and special treatment of the backlash character in random passwords, and a doubled space was removed.

Signed-off-by: MichaIng <micha@dietpi.com>