Added date format of PHP time stamp
It is very helpfull to see, according to which format the timestamp in the logs is created.

E.g. for the input plugin "tail" of the program Telegraf it is essential to parse the time given in every log entry.
To do so, a date format according to the "time" package of GO is required. 

Signed-off-by: grafjan <83698849+grafjan@users.noreply.github.com>