Fix .htaccess configuration to actually work
The `.htaccess` rules should not have the starting slash otherwise they
will never get matched. When running the mod_rewrite in .htaccess, the
path is already sanitized and leading slashes stripped.

Document also that this is not the case when the rules are used directly
from the main apache2 configuration the leading slash is present always,
which led to some edit wars of this documentation in the past as it was
not well explained (#8093, #3811).

This matches the code shipped with the .htaccess in the nextcloud server:
https://github.com/nextcloud/server/blob/master/.htaccess

Signed-off-by: Jakub Jelen <jakuje@gmail.com>