Database installation - MariaDB use mariadb / no FLUSH PRIVILEGES
MariaDB is deprecating off its mysql client usage (has been a symlink for a very long time).

FLUSH PRIVILEGES aren't required after any creation of users or grants. It has only two very specific meenings that aren't the case here; when credentials are directly changes in underlying tables using DML, or the server was started with --skip-grant-tables and this is ending that lack of priv scheme.

Signed-off-by: Daniel Black <daniel@mariadb.org>