Update Apache LimitRequestBody for >1 GiB chunking too
The recent addition clarified Apache's new behavior and and this setting's impact on non-chunking clients. With the recent v2 chunking client implementations, the maxChunkSize can be as high as 5 GiB by default (e.g. see nextcloud/desktop#6222). This expands the language to note impact for chunking clients too and clarifies the available options.

Signed-off-by: Josh Richards <josh.t.richards@gmail.com>