Update external_storage_configuration_gui.rst - details about Folder Name.
In a recent install, I discovered that the "Folder Name" field allows the users to include slashes as though creating a full path to the folder name. 

However, if you try this, the new Storage Folders malfunction and do not show up for users.

Section added explaining what the Folder Name is, what it is used for, and that it must be a single Folder Name rather than a full path.

Signed-off-by: Brent Hugh <brent@brenthugh.com>