Clarify username for postgresql's peer authentication method
The documentation is not clear that the database user name should be the name of the PHP process when using PostgreSQL's peer authentication method. This was only hinted at in a subsequent paragraph, after I had run the PSQL commands with the user name `username`.

I also updated the example to use the common `www-data` user name, which will probably familiar to most users.

Signed-off-by: Mario Campos <mario.andres.campos@gmail.com>