fix(nginx): `fastcgi_request_buffering` must be `on`
* resolves: https://github.com/nextcloud/documentation/issues/9574

Some clients like DAVx5 use chunked transfer encoding.
PHP-FPM does not support this[1] because it requires to comply with the
CGI standard - 20+ years old - which enforces a Content-Length header.

So we need to buffer the whole request until passing it to PHP-FPM.

Without this 0-byte files would be created on Nextcloud as PHP-FPM will
not only set content-length to 0 but also does not pass the body data
stream.

[1] https://bugs.php.net/bug.php?id=51191 and https://github.com/php/php-src/pull/7509

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>