Add note about davfs2 group creation
I just went through setting up davfs2 on Arch and it did not create the davfs2 group for me. I also had to edit the config to make it use the group, once I created it, because it was defaulting to network. I was getting the following error when trying to mount it:

user <username> must be member of group network

Please feel free to adjust my comment if need be. :)

Signed-off-by: Seth <github@sesu.tech>